修改类型

+++ src/main/java/com/nz/onlineMonitoring/realData/model/Meteorological.java
@@ -7,7 +7,6 @@
  **/
 package com.nz.onlineMonitoring.realData.model;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -27,35 +26,35 @@ public class Meteorological {
     /**设备编码-与t_devive dev_code关联,所属表字段为 t_meteorological.dev_code  */
     private String dev_code;
     /**空气温度,所属表字段为 t_meteorological.air_t  */
-    private BigDecimal air_t;
+    private Double air_t;
     /**空气湿度,所属表字段为 t_meteorological.air_h  */
-    private BigDecimal air_h;
+    private Double air_h;
     /**风速,所属表字段为 t_meteorological.wind_s  */
-    private BigDecimal wind_s;
+    private Double wind_s;
     /**风向,所属表字段为 t_meteorological.wind_d  */
-    private BigDecimal wind_d;
+    private Double wind_d;
     /**露点温度,所属表字段为 t_meteorological.dew_p  */
-    private BigDecimal dew_p;
+    private Double dew_p;
     /**降雨量,所属表字段为 t_meteorological.rain_f  */
-    private BigDecimal rain_f;
+    private Double rain_f;
     /**日照时数,所属表字段为 t_meteorological.sunshine_h  */
-    private BigDecimal sunshine_h;
+    private Double sunshine_h;
     /**大气压,所属表字段为 t_meteorological.atmo_s  */
-    private BigDecimal atmo_s;
+    private Double atmo_s;
     /**土壤温度1,所属表字段为 t_meteorological.soil_t1  */
-    private BigDecimal soil_t1;
+    private Double soil_t1;
     /**土壤温度2,所属表字段为 t_meteorological.soil_t2  */
-    private BigDecimal soil_t2;
+    private Double soil_t2;
     /**土壤温度3,所属表字段为 t_meteorological.soil_t3  */
-    private BigDecimal soil_t3;
+    private Double soil_t3;
     /**土壤湿度1,所属表字段为 t_meteorological.soil_h1  */
-    private BigDecimal soil_h1;
+    private Double soil_h1;
     /**土壤湿度2,所属表字段为 t_meteorological.soil_h2  */
-    private BigDecimal soil_h2;
+    private Double soil_h2;
     /**土壤湿度3,所属表字段为 t_meteorological.soil_h3  */
-    private BigDecimal soil_h3;
+    private Double soil_h3;
     /**土壤EC值,所属表字段为 t_meteorological.soil_ec  */
-    private BigDecimal soil_ec;
+    private Double soil_ec;
     /**时间,所属表字段为 t_meteorological.time  */
     private Date time;
     /**设备状态-关联表，见数据字典,所属表字段为 t_meteorological.dev_status  */
@@ -145,7 +144,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.air_t  ,空气温度
      */
-    public BigDecimal getAir_t() {
+    public Double getAir_t() {
         return air_t;
     }
 
@@ -154,7 +153,7 @@ public class Meteorological {
      *
      * @param air_t  t_meteorological.air_t,空气温度
      */
-    public void setAir_t(BigDecimal air_t) {
+    public void setAir_t(Double air_t) {
         this.air_t = air_t;
     }
 
@@ -163,7 +162,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.air_h  ,空气湿度
      */
-    public BigDecimal getAir_h() {
+    public Double getAir_h() {
         return air_h;
     }
 
@@ -172,7 +171,7 @@ public class Meteorological {
      *
      * @param air_h  t_meteorological.air_h,空气湿度
      */
-    public void setAir_h(BigDecimal air_h) {
+    public void setAir_h(Double air_h) {
         this.air_h = air_h;
     }
 
@@ -181,7 +180,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.wind_s  ,风速
      */
-    public BigDecimal getWind_s() {
+    public Double getWind_s() {
         return wind_s;
     }
 
@@ -190,7 +189,7 @@ public class Meteorological {
      *
      * @param wind_s  t_meteorological.wind_s,风速
      */
-    public void setWind_s(BigDecimal wind_s) {
+    public void setWind_s(Double wind_s) {
         this.wind_s = wind_s;
     }
 
@@ -199,7 +198,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.wind_d  ,风向
      */
-    public BigDecimal getWind_d() {
+    public Double getWind_d() {
         return wind_d;
     }
 
@@ -208,7 +207,7 @@ public class Meteorological {
      *
      * @param wind_d  t_meteorological.wind_d,风向
      */
-    public void setWind_d(BigDecimal wind_d) {
+    public void setWind_d(Double wind_d) {
         this.wind_d = wind_d;
     }
 
@@ -217,7 +216,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.dew_p  ,露点温度
      */
-    public BigDecimal getDew_p() {
+    public Double getDew_p() {
         return dew_p;
     }
 
@@ -226,7 +225,7 @@ public class Meteorological {
      *
      * @param dew_p  t_meteorological.dew_p,露点温度
      */
-    public void setDew_p(BigDecimal dew_p) {
+    public void setDew_p(Double dew_p) {
         this.dew_p = dew_p;
     }
 
@@ -235,7 +234,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.rain_f  ,降雨量
      */
-    public BigDecimal getRain_f() {
+    public Double getRain_f() {
         return rain_f;
     }
 
@@ -244,7 +243,7 @@ public class Meteorological {
      *
      * @param rain_f  t_meteorological.rain_f,降雨量
      */
-    public void setRain_f(BigDecimal rain_f) {
+    public void setRain_f(Double rain_f) {
         this.rain_f = rain_f;
     }
 
@@ -253,7 +252,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.sunshine_h  ,日照时数
      */
-    public BigDecimal getSunshine_h() {
+    public Double getSunshine_h() {
         return sunshine_h;
     }
 
@@ -262,7 +261,7 @@ public class Meteorological {
      *
      * @param sunshine_h  t_meteorological.sunshine_h,日照时数
      */
-    public void setSunshine_h(BigDecimal sunshine_h) {
+    public void setSunshine_h(Double sunshine_h) {
         this.sunshine_h = sunshine_h;
     }
 
@@ -271,7 +270,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.atmo_s  ,大气压
      */
-    public BigDecimal getAtmo_s() {
+    public Double getAtmo_s() {
         return atmo_s;
     }
 
@@ -280,7 +279,7 @@ public class Meteorological {
      *
      * @param atmo_s  t_meteorological.atmo_s,大气压
      */
-    public void setAtmo_s(BigDecimal atmo_s) {
+    public void setAtmo_s(Double atmo_s) {
         this.atmo_s = atmo_s;
     }
 
@@ -289,7 +288,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_t1  ,土壤温度1
      */
-    public BigDecimal getSoil_t1() {
+    public Double getSoil_t1() {
         return soil_t1;
     }
 
@@ -298,7 +297,7 @@ public class Meteorological {
      *
      * @param soil_t1  t_meteorological.soil_t1,土壤温度1
      */
-    public void setSoil_t1(BigDecimal soil_t1) {
+    public void setSoil_t1(Double soil_t1) {
         this.soil_t1 = soil_t1;
     }
 
@@ -307,7 +306,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_t2  ,土壤温度2
      */
-    public BigDecimal getSoil_t2() {
+    public Double getSoil_t2() {
         return soil_t2;
     }
 
@@ -316,7 +315,7 @@ public class Meteorological {
      *
      * @param soil_t2  t_meteorological.soil_t2,土壤温度2
      */
-    public void setSoil_t2(BigDecimal soil_t2) {
+    public void setSoil_t2(Double soil_t2) {
         this.soil_t2 = soil_t2;
     }
 
@@ -325,7 +324,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_t3  ,土壤温度3
      */
-    public BigDecimal getSoil_t3() {
+    public Double getSoil_t3() {
         return soil_t3;
     }
 
@@ -334,7 +333,7 @@ public class Meteorological {
      *
      * @param soil_t3  t_meteorological.soil_t3,土壤温度3
      */
-    public void setSoil_t3(BigDecimal soil_t3) {
+    public void setSoil_t3(Double soil_t3) {
         this.soil_t3 = soil_t3;
     }
 
@@ -343,7 +342,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_h1  ,土壤湿度1
      */
-    public BigDecimal getSoil_h1() {
+    public Double getSoil_h1() {
         return soil_h1;
     }
 
@@ -352,7 +351,7 @@ public class Meteorological {
      *
      * @param soil_h1  t_meteorological.soil_h1,土壤湿度1
      */
-    public void setSoil_h1(BigDecimal soil_h1) {
+    public void setSoil_h1(Double soil_h1) {
         this.soil_h1 = soil_h1;
     }
 
@@ -361,7 +360,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_h2  ,土壤湿度2
      */
-    public BigDecimal getSoil_h2() {
+    public Double getSoil_h2() {
         return soil_h2;
     }
 
@@ -370,7 +369,7 @@ public class Meteorological {
      *
      * @param soil_h2  t_meteorological.soil_h2,土壤湿度2
      */
-    public void setSoil_h2(BigDecimal soil_h2) {
+    public void setSoil_h2(Double soil_h2) {
         this.soil_h2 = soil_h2;
     }
 
@@ -379,7 +378,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_h3  ,土壤湿度3
      */
-    public BigDecimal getSoil_h3() {
+    public Double getSoil_h3() {
         return soil_h3;
     }
 
@@ -388,7 +387,7 @@ public class Meteorological {
      *
      * @param soil_h3  t_meteorological.soil_h3,土壤湿度3
      */
-    public void setSoil_h3(BigDecimal soil_h3) {
+    public void setSoil_h3(Double soil_h3) {
         this.soil_h3 = soil_h3;
     }
 
@@ -397,7 +396,7 @@ public class Meteorological {
      *
      * @return  t_meteorological.soil_ec  ,土壤EC值
      */
-    public BigDecimal getSoil_ec() {
+    public Double getSoil_ec() {
         return soil_ec;
     }
 
@@ -406,7 +405,7 @@ public class Meteorological {
      *
      * @param soil_ec  t_meteorological.soil_ec,土壤EC值
      */
-    public void setSoil_ec(BigDecimal soil_ec) {
+    public void setSoil_ec(Double soil_ec) {
         this.soil_ec = soil_ec;
     }
